fix conflict

+++ community/src/main/java/com/lawencon/community/pojo/article/PojoArticleRes.java
@@ -4,19 +4,19 @@ import java.util.List;
 
 public class PojoArticleRes {
 	private List<PojoArticleResData> data;
-	private int total;
+	private Long total;
 	public List<PojoArticleResData> getData() {
 		return data;
 	}
 	public void setData(List<PojoArticleResData> data) {
 		this.data = data;
 	}
-	public int getTotal() {
+	public Long getTotal() {
 		return total;
 	}
-	public void setTotal(int total) {
+	public void setTotal(Long total) {
 		this.total = total;
 	}
-	
+
 	
 }
